Ajout du coté admin pour le web

+++ epiforum/logic/src/main/java/com/epiforum/server/logic/dao/ProfileDao.java
@@ -14,7 +14,7 @@ public class ProfileDao {
 	@PersistenceContext(unitName="epiforum")
 	protected EntityManager	em;
 	
-	public Profile		getProfile(Long profileId) {
+	public Profile		getProfile(Integer profileId) {
 		return this.em.find(Profile.class, profileId);
 	}
